+++ server/src/Domain/Shopping/PricePolicy.cs
@@ -1,5 +1,5 @@
 namespace Metaspesa.Domain.Shopping;
 
 public static class PricePolicy {
-  public static bool IsValidPrice(float price) => price >= 0;
+  public static bool IsValidPrice(decimal price) => price >= 0;
 }
